# docs/swagger.yaml
basePath: /api
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Maintenance:
    properties:
      completed_date:
        type: string
      cost:
        type: number
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      deleted_by:
        type: integer
      id:
        type: integer
      issue_description:
        type: string
      modified_at:
        type: string
      modified_by:
        type: integer
      product_unit_id:
        type: integer
      start_date:
        type: string
      status:
        description: in_progress, completed
        type: string
      unit_code:
        type: string
    type: object
  models.MaintenanceCreateDTO:
    properties:
      cost:
        type: number
      issue_description:
        type: string
      product_unit_id:
        type: integer
      start_date:
        description: "2006-01-02"
        type: string
    required:
    - issue_description
    - product_unit_id
    - start_date
    type: object
  models.MaintenanceUpdateDTO:
    properties:
      cost:
        type: number
      issue_description:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    email: developer.haritsnb@gmail.com
    name: Harits Nala Barrun
  description: Backend RESTful API Sistem Kasir & Manajemen Rental Alat Camping menggunakan
    Golang + Gin + PostgreSQL + JWT Blacklist + KYC Upload.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Outdoor Gear Rental POS API
  version: "1.0"
paths:
...
  /maintenance:
    get:
      description: Memuat seluruh tiket servis dan perbaikan alat camping
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Maintenance'
                  type: array
              type: object
        "401":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar Semua Servis/Maintenance
      tags:
      - Maintenance
    post:
      consumes:
      - application/json
      description: Mendaftarkan unit rusak ke modul servis dan mengubah status unit
        fisik menjadi 'maintenance' (Admin/Staff Only)
      parameters:
      - description: Payload Maintenance
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MaintenanceCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Maintenance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Input Tiket Servis Manual
      tags:
      - Maintenance
  /maintenance/{id}:
    delete:
      description: Menonaktifkan catatan tiket pemeliharaan (Admin Only)
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft Delete Tiket Servis
      tags:
      - Maintenance
    get:
      description: Memuat rincian tiket perbaikan alat dan status pengerjaannya
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Maintenance'
              type: object
        "401":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail Servis Berdasarkan ID
      tags:
      - Maintenance
    post:
      description: Memulihkan kembali tiket servis yang berstatus soft-deleted (Admin
        Only)
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Restore Tiket Servis
      tags:
      - Maintenance
    put:
      consumes:
      - application/json
      description: Memperbarui deskripsi kerusakan, estimasi biaya, atau tanggal mulai
        servis (Admin/Staff Only)
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Update Maintenance
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MaintenanceUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Tiket Servis (Flexible Partial)
      tags:
      - Maintenance
  /maintenance/{id}/complete:
    patch:
      description: Menyelesaikan tiket servis dan secara otomatis mengembalikan status
        unit fisik menjadi 'available' (Admin/Staff Only)
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Selesaikan Perbaikan Unit
      tags:
      - Maintenance
  /maintenance/{id}/force:
    delete:
      description: Menghapus fisik data tiket servis secara permanen dari database
        (Admin Only)
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Permanent Tiket Servis (Force Delete)
      tags:
      - Maintenance
  /maintenance/deleted:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Maintenance'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Servis Terhapus (Trash)
      tags:
      - Maintenance
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"